django/app/serializers/api/healthcheck_serializer.py
from rest_framework import serializers

class HealthCheckSerializer(serializers.Serializer):
    status = serializers.CharField()
    version = serializers.CharField()
    example_count = serializers.CharField()
    index_example_count = serializers.CharField()
