django/requirements.txt
Django==5.1.4
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-elasticsearch-dsl==8.0
elasticsearch-dsl==8.17.0
djangorestframework==3.15.2
